refactor: 模块化拆分 main.py 并添加类型注释
main.py 从 1228 行拆分为 7 个独立模块：
- _paths: 输出路径管理
- _headers: HTTP 请求头构建
- _logging: 日志配置与刷新
- _converter: Markdown 转换器 (Obsidian/标准)
- _content: HTML 内容获取与解析
- _collection: 收藏夹处理流程
- _export: ExportContext 与文件操作

其他优化：
- favorite_ops: 提取 _safe_request/_parse_response 消除重复代码
- 全局模块添加 Python 类型注释
- 清理 get_collections.py 废弃的 __main__ 块
- 删除 main.py 末尾注释代码和模块级副作用
- 更新 README 添加新模块结构和收藏管理工具说明
