+++ src/zhihu_collections/export_all.py
@@ -1,25 +1,26 @@
-from zhihu_collections import main as main_module
+# -*- coding:utf-8 -*-
+"""知乎收藏夹 — 批量导出入口（遍历所有收藏夹）"""
+from zhihu_collections._common import load_config, parse_output_path
+from zhihu_collections._export import create_export_context
+from zhihu_collections._collection import process_single_collection
 
 
-def main():
-    config = main_module.load_config()
+def main() -> None:
+    """批量导出 config.json 中所有收藏夹"""
+    config = load_config()
     base_output_path = None
     if config.get("outputPath"):
         base_output_path = str(
-            main_module.parse_output_path(config["outputPath"], config.get("os", ""))
+            parse_output_path(config["outputPath"], config.get("os", ""))
         )
 
-    context = main_module.ExportContext(
-        base_output_path=base_output_path,
-        headers=main_module._build_page_headers(),
-        api_headers=main_module._build_api_headers(),
-        cookies=main_module.load_cookies(),
-        markdown_format=config.get("markdownFormat", "obsidian"),
+    context = create_export_context(
+        config=config, base_output_path=base_output_path
     )
 
     for c in config["zhihuUrls"]:
         print(f"========== {c['name']} ==========")
-        main_module.process_single_collection(c["name"], c["url"], context)
+        process_single_collection(c["name"], c["url"], context)
 
     print("ALL DONE")
